requirements.txt
streamlit
google-generativeai
chromadb
pymongo
python-dotenv
unstructured
pdf2image
pytesseract
tiktoken